Completed setting up new code structure splitting Kai code up into three sub-projects:
Standalone - code specific to the swing GUI part of the standalone application
Applet - code specific to the GUI part of the applet application (currently empty)
Core - contains any code that will be used by both the Applet and Standalone projects (that is where most of Kai's code is copied to)

Also the Standalone and Applet projects are setup to depend on the Core project.

Thanks to Damian for more help on using netbeans, and have a nice weekend guys.