Fixed block width/height in MovingAverage. In fact looking at it I realized no adjustment needed for block width/height here since when smoothing only the values changes not the block layout.
However, still something wrong with the block widths. This must be
when they are first created. You can see this effect when you resize
the window, where the lines appear to be moving around all over the
place. Unless there is something fundamentally wrong with JFreeChart
then the only other reason is because the block width of each data is
too large and therefore when you resize the window in the x-direction
then the overlapping block width causes the strange effect. 

One way of testing the latter is to artificially make the block width
so small that there is a gap between the points along the x-axis. Increasing
the window along the x-direction should then have no other effect than 
creasing the gap between the points.